edi_xml: improve base test class
