+++ edi_xml_oca/tests/test_xml.py
@@ -2,9 +2,8 @@
 # @author: Simone Orsi <simahawk@gmail.com>
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
-from odoo.addons.component.tests.common import SavepointComponentCase
 
-from .common import XMLTestCaseMixin
+from .common import XMLComponentTestCase
 
 TEST_XML = """<?xml version="1.0" encoding="UTF-8"?>
 <xs:element
@@ -15,17 +14,7 @@ TEST_XML = """<?xml version="1.0" encoding="UTF-8"?>
 """
 
 
-class XMLTestCase(SavepointComponentCase, XMLTestCaseMixin):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        cls.backend = cls.env.ref("edi_oca.demo_edi_backend")
-        cls.handler = cls.backend._find_component(
-            cls.backend._name,
-            ["edi.xml"],
-            work_ctx={"schema_path": "edi_xml_oca:tests/fixtures/Test.xsd"},
-        )
-
+class XMLTestCase(XMLComponentTestCase):
     def test_xml_schema_fail(self):
         with self.assertRaises(ValueError):
             self.backend._find_component(
